requirements.txt
Flask>=2.0.0
Authlib>=1.0.0
Flask-Login>=0.5.0
requests>=2.0.0
python-dotenv>=0.15.0
